+++ CHANGELOG.md
@@ -45,6 +45,8 @@ All notable changes to this project will be documented in this file.
   - `DuckDB::Result#_to_double`
   - `DuckDB::Result#_to_string`
   - `DuckDB::Result#_to_string_internal`
+  - `DuckDB::Result#_to_blob`
+  - `DuckDB::Result#to_value`
   - `DuckDB::Result#row(index)`
 - `DuckDB::Result#use_chunk_each?` is always true.
 - `DuckDB::Result#use_chunk_each=` always ignores the argument.